+++ R/storms.R
@@ -1,40 +1,52 @@
 #' Get NOAA wind storm tabular data, metadata, or shp files from IBTrACS
 #'
-#' @export
-#'
+#' @name storms
 #' @param basin (character) A basin name, one of EP, NA, NI, SA, SI, SP, or WP.
-#' @param storm (character) A storm serial number of the form YYYYJJJHTTNNN. See Details.
+#' @param storm (character) A storm serial number of the form YYYYJJJHTTNNN.
+#' See Details.
 #' @param year (numeric) One of the years from 1842 to 2014
-#' @param path (character) A path to store the files, Default: \code{~/.rnoaa/storms}
-#' @param overwrite (logical) To overwrite the path to store files in or not, Default: TRUE.
+#' @param overwrite (logical) To overwrite the path to store files in or not,
+#' Default: \code{TRUE}
 #' @param what (character) One of storm_columns or storm_names.
-#' @param type (character) One of points or lines. This gives shp files with points, or with lines.
+#' @param type (character) One of points or lines. This gives shp files with
+#' points, or with lines.
 #' @param x Output from \code{storm_shp}, a path to shp file to read in.
 #' @param ... Curl options passed on to \code{\link[httr]{GET}} (optional)
+#'
+#' @return A tibble (data.frame)
 #'
 #' @details International Best Track Archive for Climate Stewardship (IBTrACS)
 #'
 #' Details for storm serial numbers:
 #' \itemize{
-#'  \item YYYY is the corresponding year of the first recorded observation of the storm
+#'  \item YYYY is the corresponding year of the first recorded observation of
+#'  the storm
 #'  \item JJJ is the day of year of the first recorded observation of the storm
 #'  \item H is the hemisphere of the storm: N=Northern, S=Southern
-#'  \item TT is the absolute value of the rounded latitude of the first recorded observation of the
-#'  storm (range 0-90, if basin=SA or SH, then TT in reality is negative)
-#'  \item NNN is the rounded longitude of the first recorded observation of the storm (range 0-359)
+#'  \item TT is the absolute value of the rounded latitude of the first
+#'  recorded observation of the storm (range 0-90, if basin=SA or SH, then TT
+#'  in reality is negative)
+#'  \item NNN is the rounded longitude of the first recorded observation of
+#'  the storm (range 0-359)
 #' }
 #'
-#' For example: \code{1970143N19091} is a storm in the North Atlantic which started on
-#' May 23, 1970 near 19N 91E
+#' For example: \code{1970143N19091} is a storm in the North Atlantic which
+#' started on May 23, 1970 near 19N 91E
 #'
-#' See \url{http://www.ncdc.noaa.gov/ibtracs/index.php?name=numbering} for more.
+#' See \url{http://www.ncdc.noaa.gov/ibtracs/index.php?name=numbering} for more
 #'
 #' The datasets included in the package \code{\link{storm_names}}, and
 #' \code{\link{storm_columns}} may help in using these storm functions.
 #'
+#'
+#' @section File storage:
+#' We use \pkg{rappdirs} to store files, see
+#' \code{\link[rappdirs]{user_cache_dir}} for how
+#' we determine the directory on your machine to save files to, and run
+#' \code{user_cache_dir("rnoaa")} to get that directory.
+#'
 #' @references \url{http://www.ncdc.noaa.gov/ibtracs/index.php?name=wmo-data}
 #'
-#' @rdname storms
 #' @examples \dontrun{
 #' # Metadata
 #' head( storm_meta() )
@@ -77,34 +89,40 @@
 #' plot(res4shp)
 #' }
 
-storm_data <- function(basin=NULL, storm=NULL, year=NULL, path="~/.rnoaa/storms",
-                       overwrite = TRUE, ...){
+#' @export
+#' @rdname storms
+storm_data <- function(basin = NULL, storm = NULL, year = NULL,
+                       overwrite = TRUE, ...) {
+  calls <- names(sapply(match.call(), deparse))[-1]
+  calls_vec <- "path" %in% calls
+  if (any(calls_vec)) {
+    stop("The parameter path has been removed, see ?storms",
+         call. = FALSE)
+  }
 
+  path <- file.path(rnoaa_cache_dir, "storms")
   csvpath <- csv_local(basin, storm, year, path)
   if (!is_storm(x = csvpath)) {
     csvpath <- storm_GET(path, basin, storm, year, overwrite, ...)
   }
   message(sprintf("<path>%s", csvpath), "\n")
-  structure(list(data = read_csv(csvpath)), class = "storm_data")
-}
-
-#' @export
-print.storm_data <- function(x, ..., n = 10){
-  cat("<NOAA Storm Data>", sep = "\n")
-  cat(sprintf("Size: %s X %s\n", NROW(x$data), NCOL(x$data)), sep = "\n")
-  trunc_mat_(x$data, n = n)
+  tibble::as_data_frame(read_csv(csvpath))
 }
 
 storm_GET <- function(bp, basin, storm, year, overwrite, ...){
-  dir.create(local_base(basin, storm, year, bp), showWarnings = FALSE, recursive = TRUE)
+  dir.create(local_base(basin, storm, year, bp), showWarnings = FALSE,
+             recursive = TRUE)
   fp <- csv_local(basin, storm, year, bp)
-  res <- suppressWarnings(GET(csv_remote(basin, storm, year), write_disk(fp, overwrite), ...))
+  res <- suppressWarnings(GET(csv_remote(basin, storm, year),
+                              write_disk(fp, overwrite), ...))
   res$request$output$path
 }
 
 filecheck <- function(basin, storm, year){
   tmp <- noaa_compact(list(basin = basin, storm = storm, year = year))
-  if (length(tmp) > 1) stop("You can only supply one or more of basin, storm, or year")
+  if (length(tmp) > 1)
+    stop("You can only supply one or more of basin, storm, or year",
+         call. = FALSE)
   if (length(tmp) == 0) list(all = "Allstorms") else tmp
 }
 
@@ -120,11 +138,19 @@ filepath <- function(basin, storm, year){
 
 fileext <- function(basin, storm, year){
   tt <- filepath(basin, storm, year)
-  if (grepl("Allstorms", tt)) paste0(tt, '.ibtracs_all.v03r06.csv.gz') else paste0(tt, '.ibtracs_all.v03r06.csv')
+  if (grepl("Allstorms", tt)) {
+    paste0(tt, '.ibtracs_all.v03r06.csv.gz')
+  } else {
+    paste0(tt, '.ibtracs_all.v03r06.csv')
+  }
 }
 
-csv_remote <- function(basin, storm, year) file.path(stormurl(), fileext(basin, storm, year))
-csv_local <- function(basin, storm, year, path) file.path(path, fileext(basin, storm, year))
+csv_remote <- function(basin, storm, year) {
+  file.path(stormurl(), fileext(basin, storm, year))
+}
+csv_local <- function(basin, storm, year, path) {
+  file.path(path, fileext(basin, storm, year))
+}
 
 local_base <- function(basin, storm, year, path){
   tt <- filecheck(basin, storm, year)
@@ -133,4 +159,6 @@ local_base <- function(basin, storm, year, path){
 
 is_storm <- function(x) if (file.exists(x)) TRUE else FALSE
 
-stormurl <- function(x = "csv") sprintf('ftp://eclipse.ncdc.noaa.gov/pub/ibtracs/v03r06/all/%s', x)
+stormurl <- function(x = "csv") {
+  sprintf('ftp://eclipse.ncdc.noaa.gov/pub/ibtracs/v03r06/all/%s', x)
+}