+++ R/storms_meta.R
@@ -1,6 +1,5 @@
 #' @export
 #' @rdname storms
-storm_meta <- function(what="storm_columns")
-{
+storm_meta <- function(what = "storm_columns") {
   switch(what, storm_columns = storm_columns, storm_names = storm_names)
 }