+++ tests/testthat/test-seaice.R
@@ -4,7 +4,7 @@ test_that("seaice functions work", {
   skip_on_cran()
   skip_on_travis()
 
-  url <- seaiceeurls(1980, mo = 'Feb', pole = 'S')
+  url <- seaiceeurls(yr = 1980, mo = 'Feb', pole = 'S')
   out <- seaice(url)
 
   expect_is(url, "character")